Data.Conduit.Network.Internal

Documentation

data AppData m

The data passed to an Application.

Since 0.6.0

Constructors

AppData

Fields

appSource :: Source m ByteString

appSink :: Sink ByteString m ()

appSockAddr :: SockAddr

appLocalAddr :: Maybe SockAddr

data ServerSettings m

Settings for a TCP server. It takes a port to listen on, and an optional hostname to bind to.

Since 0.6.0

Constructors

ServerSettings

Fields

serverPort :: Int

serverHost :: HostPreference

serverAfterBind :: Socket -> m ()

serverNeedLocalAddr :: Bool

data ClientSettings m

Settings for a TCP client, specifying how to connect to the server.

Since 0.6.0

Constructors

ClientSettings

Fields

clientPort :: Int

clientHost :: ByteString