Data.Conduit.Network.UDP

UDP message representation

data Message

Representation of a single message

Constructors

Message

Fields

msgData :: !ByteString

msgSender :: !SockAddr

Basic utilities

sourceSocket :: MonadIO m => Socket -> Int -> GSource m Message

Stream messages from the socket.

The given len defines the maximum packet size. Every produced item contains the message payload and the origin address.

This function does not automatically close the socket.

sinkSocket :: MonadIO m => Socket -> GInfSink ByteString m

Stream messages to the connected socket.

The payload is sent using send, so some of it might be lost.

This function does not automatically close the socket.

sinkAllSocket :: MonadIO m => Socket -> GInfSink ByteString m

Stream messages to the connected socket.

The payload is sent using sendAll, so it might end up in multiple packets.

This function does not automatically close the socket.

sinkToSocket :: MonadIO m => Socket -> GInfSink Message m

Stream messages to the socket.

Every handled item contains the message payload and the destination address. The payload is sent using sendTo, so some of it might be lost.

This function does not automatically close the socket.

sinkAllToSocket :: MonadIO m => Socket -> GInfSink Message m

Stream messages to the socket.

Every handled item contains the message payload and the destination address. The payload is sent using sendAllTo, so it might end up in multiple packets.

This function does not automatically close the socket.

Helper Utilities

data HostPreference

Which host to bind.

Note: The IsString instance recognizes the following special values:

* means HostAny

*4 means HostIPv4

*6 means HostIPv6

Constructors

HostAny

HostIPv4

HostIPv6

Host String

bindPort :: Int -> HostPreference -> IO Socket

Attempt to bind a listening Socket on the given host/port. If no host is given, will use the first address available.

getSocket :: String -> Int -> IO (Socket, AddrInfo)

Attempt to connect to the given host/port.